+++ changelog.md
@@ -16,6 +16,7 @@
 - Fixed bug where "default" variants of Forestry leaves would never be used
 - Fixed bug where fruit ripeness would sometimes not persist
 - ACTUALLY make saplings and other items usable in the composter (#35)
+- Increase energy storage of Bottler, Still, Fermenter by 10x
 
 ## Forestry 1.0.4
 - Added modify genome command. Can be used like `/forestry bee modify <chromosome> <allele> <both|dominant|recessive>` to